Rework release API (releaseMono, asynchronous imperative release)
The preferred way of releasing is now `void release`, which doesn't
block but fires and forgets. The underlying Mono is still accessible
through `releaseMono`, for the cases where it is necessary to wait for
the end of the release-and-recycle phase. These use cases are expected
to be the exception rather than the norm.