Big refactor of API/implementation to add intermediate PoolSlot
This commit introduces PoolSlot, an intermediate abstraction that
sits between the Pool and POOLABLE objects which allows to keep
track of usage (for statistics and, later, eviction purposes).

As a result the Pool API now only deals with borrowing objects, and
release is performed at the PoolSlot level.

In order to avoid dealing with PoolSlot directly (with the constraint
of having to keep a reference to the PoolSlot for subsequent
release), there are now two methods of borrowing: acquire() exposes
the PoolSlot, while borrow(Function) wraps the PoolSlot in a
Mono.usingWhen, forcing the user to be descriptive about her usage
of the object and allowing for automatically performing release.