WIP rework of AffinityPool, use abstract classes and redo whole implem
The new implementation works around the idea that borrowing/releasing
from a same Thread will lead to direct exchange of the PooledRef, as
long as there is a thread-local pending Borrower.

As soon as this is not the case, we enter a slow path loop where a
single "starved" thread will win, looking at subpools that are not
currently in a fast path for pending Borrowers.