Rework builder, introduce AllocationStrategy, make pool implems public
The PoolBuilder has been simplified into a PoolConfigBuilder, with less
intermediate step interfaces (as most elements are really optional).

It has been moved back in the `api` package and creates a PoolConfig
rather than Pools directly. The `builder` package has been renamed
`impl`.

Pool implementations are now public and created by simply passing in a
PoolConfig.

The minSize/maxSize has been removed in favor of an AllocationStrategy,
which atomically maintains the state that was previously managed in
the pools (LIVE volatile and AtomicIntegerFieldUpdater).