[fix] Rework slowPathRecycle to defer polling available elements
Instead of polling the availableElements queue to re-offer the element
when no pending is found, defer the polling until after a pending is
found (use a peek upfront).

This _could_ result in race with acquire / polling a null when peeking
a non-null, in which case we re-enqueue the "borrower". This means the
pending borrower _can_ be reordered.