fix #47 Add Pool#disposeLater for async disposal
This commit introduces an async API for pool disposal/shutdown, via the
`disposeLater` method.

The `dispose()` method is now a default interface method that relies on
said `disposeLater()` API, by simply subscribing to its returned Mono.

`disposeLater()` Mono is cold, only triggering shutdown on _first_
subscription. It can be used to wait for the full release of pooled
resources, eg. by `block()`ing on it.