fix #51 Expose PoolConfig, rework PoolBuilder with factory (#55)
This commit heavily reworks how the PoolBuilder is producing
a PoolConfig (internally) and then a concrete Pool instance.

The goal is to open the builder for cases where implementations
would be provided by external projects (eg. a netty event-loop
aware implementation for a project that pools netty connections).

One can pass a factory Function to the builder to create such
a custom pool, possibly also switching the builder to a custom
type of configuration beforehand. As a result, the
DefaultPoolConfig has been made public, extracting a PoolConfig
interface.

As a shorthand for the `build(Function)` that produces
reactor-pool vanilla implementations, we now expose `lifo()` and
`fifo()`, which are now the two out-of-the-box provided flavors.
These are explicitly marked as returning an `InstrumentedPool`.