fix #53 Remove initSize, add permitMinimum and do warmup on acquire
This commit lets `AllocationStrategy` answer `getPermits(0)` with warmup
size, and more generally answer `getPermits(n)` with a number greater
than n. This represents the actual number of resources to allocate in
the pool in response to the initial permit request, which can typically
be greater than the request when there is a need to maintain a minimum
of live resources (warmup).

The notion of a minimum size is thus introduced, replacing the builder's
`initSize` concept with that of `sizeMin` and `sizeBetween`.

The `SimplePool` uses this minimum to trigger a re-warmup of resources
whenever it allocates missing resources during `acquire()`, so as to
maintain this floor of resources.
In constructor, the loop to allocate `initSize` resources is replaced
by a call to the same (internal) warmup method.