fix #5 Add warmup() API and remove implicit constructor warmup
The warmup API uses the `AllocationStrategy` to "warm up" a minimal set
of resources manually.

SimplePool constructor no longer allocates resources, so user MUST now
explicitly call `warmup()` upon obtaining the pool if they want to wait
for the minimal set to be ready and willing to pay that cost upfront.

Alternatively, this warmup will be implicitly done at the first call to
`acquire()`. Note however that `acquire()` latency shouldn't be affected
by the warmup, provided allocations are non-blocking.