fix #25 Make each acquirer get a different limited PooledRef instance
A new PooledRef is put back in the pool each time the previous one is
released. This in turns allows to put a guard in place on release() and
invalidate() in order to avoid multiple subscriptions to the resulting
Monos to trigger actually releasing the underlying resource more than
once.