fix #60 Separate Clock and PoolMetricsRecorder
This commit removes the responsibility of timestamping from the
PoolMetricsRecorder and puts it on a Clock instance that can be passed
to the PoolBuilder.

`now()` and `measureTime(long)` have been removed from the former, and
pool implementations now use the configured Clock instead. The default
clock is `Clock.SystemUTC()` (since only the millis() method is used).