Be more strict with permit return
This commit checks that for the bounded allocation strategy, the number
of returned permits is consistent with the max. If the returned number
+ current remaining permit count would pass the max, `returnPermits` now
throws an IllegalArgumentException (which is a possibility documented on
the AllocationStrategy's javadoc).

This commit changes one test that would bypass getPermits and thus end
up with too many permits.