Let invalidate/release run on IDLE state, with a CAS loop
This commit let invalidate() and release() run if the state of
the PooledRef is still IDLE. Furthermore, it retries updates in a
CAS loop to ensure no concurrent release/invalidate was executed.

The `markReleased()` method now returns a boolean indicating if it
was able to perform the marking, which can be used in the normal
path to avoid release-related accounting when not necessary.

Note that a slot in IDLE state will (correctly) be released/invalidated
with this change (transiting to RELEASED is no longer limited to the
ACQUIRED state).