fix #84 Prevent warmup from breaking the FIFO order
This commit reworks the warmup code, enforcing it to run after the
main allocation, so that if there is a transient error it doesn't break
the FIFO/LIFO order.

Additionally it ensures that ACQUIRED is only decremented by the main
allocation, not the warmup. Permits are returned correctly as each of
the warmup fails.