fix #86 Add background eviction
This commit adds an option for background eviction of resources in case
no activity is registered on the pool.

Note that the background eviction is best effort, and will back off
if there are concurrent release / acquire activity. The later actually
checks the eviction predicate too, which allows for some amount of
eviction to occur even if the background task has backed off.