Add SamplingAllocationStrategy, in new introspection util package (#113)
This commit adds a SamplingAllocationStrategy in the newly introduced
introspection package. It allows to sample calls to the AllocationStrategy that
is wrapped, effectively allowing to trace a percentage of these calls.

getPermits and returnPermits can be sampled at separate rates, and the samples
can be inspected directly (as public final fields).