Polish ParameterizedTest reports (#120)
This commit polishes the tests, and notably how ParameterizedTest are
reported in console/CI. Currently, the method names are omitted which
makes parsing the logs for eg. CommonPoolTest very complicated. This
is because the default name pattern fo a ParameterizedTest only
includes the index and parameter values (sic). By adding the
{displayName} placeholder (explicitly or via the meta-annotation we
introduce), we get the method name by default.

 - Bump JUnit 5.6.0 => 5.7.1
 - Nested test class can now be static
 - Add ParameterizedTestWithName meta-annotation, ensure test method
   always shown
 - Ensure Nested tests are shown in Gradle output, removed their
   displayName