GitHub Actions Release Process: signing, artifactory, sonatype
This commit changes the release process to be based on GitHub Actions.

CI is already performed on GHA, but that workflow will now only run
for pull_request trigger.

We define helper tasks and methods in the setup.gradle file to help
interpret the version number. The trigger is pushing a commit to an
active maintenance branch.

The new workflow uses GitHub Environments to manage access to secrets,
especially in the case of milestones and releases as it introduces
signing of artifacts in these cases. It is split into 4 jobs, 3 of
which are conditional and depend on a GitHub Environment.

Notably:
 - Add a function to qualify a version number as one of RELEASE,
 MILESTONE, SNAPSHOT or BAD
 - Add a GitHub Actions oriented task, `qualifyVersionGha`, which
 produces gha step output for the current project version. `versionType`
 is an application of the above function and `fullVersion` is the plain
 project's `version`
 - tag releases and milestones using `git` commands in `prepare` job
 - Activate the JFrog Artifactory plugin only if a password is defined
 (including in subprojects)
 - Add the `signing` plugin, activated only if a signing key is defined,
   requiring that signing be performed if a publish task is invoked
 - Add a Maven publication repository if Sonatype OSS credentials are
   defined
 - Add a `mock` maven repository target and a `-PforceSigning` option
 for debugging purposes
 - Use the GHA RUN_NUMBER as the build id in Artifactory

See reactor/reactor#694.
See reactor/reactor-addons#254.