[chores] Add Spotless plugin for incremental bump in copyright range
This commit introduces the Spotless plugin, which is capable of only
applying checks to files that have changed compared to a given branch.

Replacing hyeronimus license plugin, this ensures that in the future the
license headers will have the year range updated whenever a file is
touched (and not as a big bang operation on Jan 1st, like with the old
plugin).

See reactor/reactor#701