Prevent invalidate() usage on a release() ref, fix acquireCount
This commit improves the guarding of calling #invalidate() on a ref that
has already been released.

It prevents the acquireCount metric to go below zero in such a case.

The current state of the codebase would make it difficult to allow an
invalidate() call on the same ref #release() was called on.
When an invalidate() would have been desirable post-release, the
background eviction feature should suffice.

However, this aspect was not made that explicit in the javadoc, which
has been amended to better reflect the constraints between the various
PooledRef methods.

See #146.