Introduce decorators, add GracefulShutdown feature as decorator (#149)
This commit introduces `InstrumentedPoolDecorators`, as well as
`PoolBuilder#buildPoolAndDecorateWith` helper in the build for
directly applying decorators.

The first such decorator adds the feature to "gracefully shutdown"
any `InstrumentedPool`.

For such a decorated pool, calling `disposeGracefully(Duration)`
will start a countdown for the given duration. During that time,
no new `acquire` will be accepted. However, pending acquires and
currently acquired resources will be given a chance to be utilized
and released back to the pool. Whenever all pending and currently
acquired resources are released OR when the timeout expires, the
pool will finally be disposed.

Fixes #148.